third_party/blink/web_tests/platform/mac-mac10.10/virtual/not-omt-sw-fetch/external/wpt/xhr/getallresponseheaders-expected.txt

This is a testharness.js-based test.
FAIL XMLHttpRequest: getAllResponseHeaders() assert_equals: expected "also-here: Mr. PB\r\newok: lego\r\nfoo-test: 1, 2\r\n__custom: token\r\n" but got "__custom: token\r\nalso-here: Mr. PB\r\nfoo-test: 1, 2\r\newok: lego\r\n"
PASS XMLHttpRequest: getAllResponseHeaders() 1
PASS XMLHttpRequest: getAllResponseHeaders() 2
PASS XMLHttpRequest: getAllResponseHeaders() 3
PASS XMLHttpRequest: getAllResponseHeaders() 4
PASS XMLHttpRequest: getAllResponseHeaders() 5
PASS XMLHttpRequest: getAllResponseHeaders() 6
Harness: the test ran to completion.
